remove levelupmodal, more rpg unsnarling

+++ Assets/CommonCoreGame/RPGGame/RPG/RpgDefaultValues.cs
@@ -18,8 +18,8 @@ namespace CommonCore.RpgGame.Rpg
         public static int PotentialPointsForLevel(int newLevel, CharacterModel character)
         {
             return (2
-                + (int)(0.5f * (float)character.BaseStats.Stats[StatType.Erudition])
-                + (int)(0.25f * (float)character.BaseStats.Stats[StatType.Intuition])
+                + (int)(0.5f * (float)character.BaseStats.Stats[(int)StatType.Erudition])
+                + (int)(0.25f * (float)character.BaseStats.Stats[(int)StatType.Intuition])
                 );
         }
 
@@ -62,19 +62,19 @@ namespace CommonCore.RpgGame.Rpg
         public static float MaxEnergy(CharacterModel characterModel)
         {
             return Mathf.FloorToInt(100
-                + characterModel.DerivedStats.Stats[StatType.Dexterity] * 5.0f
-                + characterModel.DerivedStats.Stats[StatType.Resilience] * 10.0f
-                + characterModel.DerivedStats.Skills[SkillType.Athletics] * 0.5f
-                + characterModel.DerivedStats.Skills[SkillType.AthleticsFleet] * 0.25f);
+                + characterModel.DerivedStats.Stats[(int)StatType.Dexterity] * 5.0f
+                + characterModel.DerivedStats.Stats[(int)StatType.Resilience] * 10.0f
+                + characterModel.DerivedStats.Skills[(int)SkillType.Athletics] * 0.5f
+                + characterModel.DerivedStats.Skills[(int)SkillType.AthleticsFleet] * 0.25f);
         }
 
         public static float MaxMagic(CharacterModel characterModel)
         {
             return Mathf.FloorToInt(100
-                + characterModel.DerivedStats.Stats[StatType.Erudition] * 5.0f
-                + characterModel.DerivedStats.Stats[StatType.Intuition] * 5.0f
-                + characterModel.DerivedStats.Stats[StatType.Serendipity] * 5.0f
-                + characterModel.DerivedStats.Skills[SkillType.Magic] * 0.5f);
+                + characterModel.DerivedStats.Stats[(int)StatType.Erudition] * 5.0f
+                + characterModel.DerivedStats.Stats[(int)StatType.Intuition] * 5.0f
+                + characterModel.DerivedStats.Stats[(int)StatType.Serendipity] * 5.0f
+                + characterModel.DerivedStats.Skills[(int)SkillType.Magic] * 0.5f);
         }
 
 
@@ -108,11 +108,11 @@ namespace CommonCore.RpgGame.Rpg
         public static int AdjustedBuyPrice(CharacterModel character, float value)
         {
             float adjValue = (value * 2.0f)
-                - (character.DerivedStats.Skills[SkillType.Social] / 200 * value)
-                - (character.DerivedStats.Skills[SkillType.SocialExchange] / 100 * value)
-                - (character.DerivedStats.Skills[SkillType.SocialLeverage] / 200 * value)
-                - (character.DerivedStats.Stats[StatType.Dialectic] / 1000 * value)
-                - (character.DerivedStats.Stats[StatType.Subterfuge] / 1000 * value);
+                - (character.DerivedStats.Skills[(int)SkillType.Social] / 200 * value)
+                - (character.DerivedStats.Skills[(int)SkillType.SocialExchange] / 100 * value)
+                - (character.DerivedStats.Skills[(int)SkillType.SocialLeverage] / 200 * value)
+                - (character.DerivedStats.Stats[(int)StatType.Dialectic] / 1000 * value)
+                - (character.DerivedStats.Stats[(int)StatType.Subterfuge] / 1000 * value);
 
             return Mathf.CeilToInt(Mathf.Max(value, adjValue));
         }
@@ -122,11 +122,11 @@ namespace CommonCore.RpgGame.Rpg
             float halfValue = value * 0.5f;
 
             float adjValue = halfValue
-                + (character.DerivedStats.Skills[SkillType.Social] / 200 * halfValue)
-                + (character.DerivedStats.Skills[SkillType.SocialExchange] / 200 * halfValue)
-                + (character.DerivedStats.Skills[SkillType.SocialLeverage] / 200 * halfValue)
-                + (character.DerivedStats.Stats[StatType.Dialectic] / 1000 * halfValue)
-                + (character.DerivedStats.Stats[StatType.Subterfuge] / 1000 * halfValue);
+                + (character.DerivedStats.Skills[(int)SkillType.Social] / 200 * halfValue)
+                + (character.DerivedStats.Skills[(int)SkillType.SocialExchange] / 200 * halfValue)
+                + (character.DerivedStats.Skills[(int)SkillType.SocialLeverage] / 200 * halfValue)
+                + (character.DerivedStats.Stats[(int)StatType.Dialectic] / 1000 * halfValue)
+                + (character.DerivedStats.Stats[(int)StatType.Subterfuge] / 1000 * halfValue);
 
             return Mathf.FloorToInt(Mathf.Min(adjValue, value));
         }
@@ -234,9 +234,9 @@ namespace CommonCore.RpgGame.Rpg
         public static float DetectionChance(CharacterModel character, bool isSneaking, bool isRunning)
         {
 
-            float r1 = character.DerivedStats.Skills[SkillType.AthleticsFurtive];
-            float r2 = character.DerivedStats.Skills[SkillType.Athletics] * 0.25f;
-            float r3 = character.DerivedStats.Stats[StatType.Dexterity] * 2f;
+            float r1 = character.DerivedStats.Skills[(int)SkillType.AthleticsFurtive];
+            float r2 = character.DerivedStats.Skills[(int)SkillType.Athletics] * 0.25f;
+            float r3 = character.DerivedStats.Stats[(int)StatType.Dexterity] * 2f;
             float rawChance = (100f - (r1 + r2 + r3)) / 100f;
 
             rawChance *= isSneaking ? 0.5f : 1.0f;
@@ -250,9 +250,9 @@ namespace CommonCore.RpgGame.Rpg
         public static float GetMoveSpeedMultiplier(CharacterModel character)
         {
             float rawSpeed = 1f
-                + (character.DerivedStats.Stats[StatType.Dexterity] / 50f)
-                + (character.DerivedStats.Skills[SkillType.AthleticsFleet] / 150f)
-                + (character.DerivedStats.Skills[SkillType.Athletics] / 300f);
+                + (character.DerivedStats.Stats[(int)StatType.Dexterity] / 50f)
+                + (character.DerivedStats.Skills[(int)SkillType.AthleticsFleet] / 150f)
+                + (character.DerivedStats.Skills[(int)SkillType.Athletics] / 300f);
             return Mathf.Clamp(rawSpeed, 0.75f, 1.75f);
         }
 
@@ -264,8 +264,8 @@ namespace CommonCore.RpgGame.Rpg
         public static float GetRunEnergyRate(CharacterModel character)
         {
             float rawRate = 1f
-                - (character.DerivedStats.Skills[SkillType.AthleticsFleet] / 500f)
-                - (character.DerivedStats.Skills[SkillType.Athletics] / 500f);
+                - (character.DerivedStats.Skills[(int)SkillType.AthleticsFleet] / 500f)
+                - (character.DerivedStats.Skills[(int)SkillType.Athletics] / 500f);
 
             return Mathf.Min(0.1f, rawRate);
         }
@@ -273,9 +273,9 @@ namespace CommonCore.RpgGame.Rpg
         public static float GetJumpVelocityMultiplier(CharacterModel character)
         {
             float rawMultiplier = 1f
-                + (character.DerivedStats.Stats[StatType.Dexterity] / 100f)
-                + (character.DerivedStats.Skills[SkillType.AthleticsFleet] / 200f)
-                + (character.DerivedStats.Skills[SkillType.Athletics] / 500f);
+                + (character.DerivedStats.Stats[(int)StatType.Dexterity] / 100f)
+                + (character.DerivedStats.Skills[(int)SkillType.AthleticsFleet] / 200f)
+                + (character.DerivedStats.Skills[(int)SkillType.Athletics] / 500f);
 
             return Mathf.Clamp(rawMultiplier, 0.75f, 1.5f);
         }
@@ -283,7 +283,7 @@ namespace CommonCore.RpgGame.Rpg
         public static float GetJumpEnergyUse(CharacterModel character)
         {
             return 10f
-                - (character.DerivedStats.Skills[SkillType.Athletics] / 50f); //athletics very slightly reduces energy use
+                - (character.DerivedStats.Skills[(int)SkillType.Athletics] / 50f); //athletics very slightly reduces energy use
         }
 
         public static float GetAirMoveMultiplier(CharacterModel character)
@@ -294,9 +294,9 @@ namespace CommonCore.RpgGame.Rpg
         public static float GetIdleEnergyRecoveryRate(CharacterModel character)
         {
             return 5f
-                + (character.DerivedStats.Stats[StatType.Dexterity] / 2f)
-                + (character.DerivedStats.Stats[StatType.Resilience] / 5f)
-                + (character.DerivedStats.Skills[SkillType.Athletics] / 100f);
+                + (character.DerivedStats.Stats[(int)StatType.Dexterity] / 2f)
+                + (character.DerivedStats.Stats[(int)StatType.Resilience] / 5f)
+                + (character.DerivedStats.Skills[(int)SkillType.Athletics] / 100f);
         }
 
         public static float GetMovingEnergyRecoveryRate(CharacterModel character)
@@ -311,10 +311,10 @@ namespace CommonCore.RpgGame.Rpg
             //higher is better
             return Mathf.Clamp(
                     1.0f
-                    + (character.DerivedStats.Stats[StatType.Resilience] / 20f)
-                    + (character.DerivedStats.Stats[StatType.Dexterity] / 30f)
-                    + (character.DerivedStats.Skills[SkillType.MeleeBrawn] / 100f)
-                    + (character.DerivedStats.Skills[SkillType.Melee] / 200f),
+                    + (character.DerivedStats.Stats[(int)StatType.Resilience] / 20f)
+                    + (character.DerivedStats.Stats[(int)StatType.Dexterity] / 30f)
+                    + (character.DerivedStats.Skills[(int)SkillType.MeleeBrawn] / 100f)
+                    + (character.DerivedStats.Skills[(int)SkillType.Melee] / 200f),
                 0.5f, 3.0f);
         }
 
@@ -323,10 +323,10 @@ namespace CommonCore.RpgGame.Rpg
             //higher is better
             return Mathf.Clamp(
                     1.0f
-                    + (character.DerivedStats.Stats[StatType.Resilience] / 20f)
-                    + (character.DerivedStats.Stats[StatType.Dexterity] / 30f)
-                    + (character.DerivedStats.Skills[SkillType.MeleeBrawn] / 100f)
-                    + (character.DerivedStats.Skills[SkillType.Melee] / 200f),
+                    + (character.DerivedStats.Stats[(int)StatType.Resilience] / 20f)
+                    + (character.DerivedStats.Stats[(int)StatType.Dexterity] / 30f)
+                    + (character.DerivedStats.Skills[(int)SkillType.MeleeBrawn] / 100f)
+                    + (character.DerivedStats.Skills[(int)SkillType.Melee] / 200f),
                 0.5f, 2.0f);
         }
 
@@ -335,10 +335,10 @@ namespace CommonCore.RpgGame.Rpg
             //lower is better
             return Mathf.Clamp(
                     1.0f
-                    - (character.DerivedStats.Stats[StatType.Resilience] / 20f)
-                    - (character.DerivedStats.Stats[StatType.Dexterity] / 30f)
-                    - (character.DerivedStats.Skills[SkillType.MeleeBrawn] / 100f)
-                    - (character.DerivedStats.Skills[SkillType.Melee] / 200f),
+                    - (character.DerivedStats.Stats[(int)StatType.Resilience] / 20f)
+                    - (character.DerivedStats.Stats[(int)StatType.Dexterity] / 30f)
+                    - (character.DerivedStats.Skills[(int)SkillType.MeleeBrawn] / 100f)
+                    - (character.DerivedStats.Skills[(int)SkillType.Melee] / 200f),
                 0.5f, 2.0f);
         }
 
@@ -350,16 +350,16 @@ namespace CommonCore.RpgGame.Rpg
             switch ((SkillType)itemModel.SkillType)
             {
                 case SkillType.Melee:
-                    factor -= character.DerivedStats.Skills[SkillType.MeleeAlacrity] / 100f;
-                    factor -= character.DerivedStats.Stats[StatType.Dexterity] / 100f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.MeleeAlacrity] / 100f;
+                    factor -= character.DerivedStats.Stats[(int)StatType.Dexterity] / 100f;
                     break;
                 case SkillType.Archery:
-                    factor -= character.DerivedStats.Skills[SkillType.Archery] / 400f; //you can't really increase archery fire rate much
-                    factor -= character.DerivedStats.Stats[StatType.Dexterity] / 100f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.Archery] / 400f; //you can't really increase archery fire rate much
+                    factor -= character.DerivedStats.Stats[(int)StatType.Dexterity] / 100f;
                     break;
                 case SkillType.Guns:
-                    factor -= character.DerivedStats.Skills[SkillType.GunsRapidity] / 100f;
-                    factor -= character.DerivedStats.Stats[StatType.Dexterity] / 200f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.GunsRapidity] / 100f;
+                    factor -= character.DerivedStats.Stats[(int)StatType.Dexterity] / 200f;
                     break;
             }
 
@@ -374,8 +374,8 @@ namespace CommonCore.RpgGame.Rpg
             switch ((SkillType)itemModel.SkillType)
             {
                 case SkillType.Guns:
-                    factor -= character.DerivedStats.Skills[SkillType.GunsRapidity] / 100f;
-                    factor -= character.DerivedStats.Stats[StatType.Dexterity] / 100f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.GunsRapidity] / 100f;
+                    factor -= character.DerivedStats.Stats[(int)StatType.Dexterity] / 100f;
                     break;
             }
 
@@ -390,12 +390,12 @@ namespace CommonCore.RpgGame.Rpg
             switch ((SkillType)itemModel.SkillType)
             {
                 case SkillType.Archery:
-                    factor -= character.DerivedStats.Skills[SkillType.ArcherySteady] / 100f;
-                    factor -= character.DerivedStats.Stats[StatType.Dexterity] / 100f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.ArcherySteady] / 100f;
+                    factor -= character.DerivedStats.Stats[(int)StatType.Dexterity] / 100f;
                     break;
                 case SkillType.Guns:
-                    factor -= character.DerivedStats.Skills[SkillType.GunsAccuracy] / 100f;
-                    factor -= character.DerivedStats.Stats[StatType.Dexterity] / 100f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.GunsAccuracy] / 100f;
+                    factor -= character.DerivedStats.Stats[(int)StatType.Dexterity] / 100f;
                     break;
             }
 
@@ -410,12 +410,12 @@ namespace CommonCore.RpgGame.Rpg
             switch ((SkillType)itemModel.SkillType)
             {
                 case SkillType.Archery:
-                    factor -= character.DerivedStats.Skills[SkillType.ArcherySteady] / 100f;
-                    factor -= character.DerivedStats.Stats[StatType.Dexterity] / 100f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.ArcherySteady] / 100f;
+                    factor -= character.DerivedStats.Stats[(int)StatType.Dexterity] / 100f;
                     break;
                 case SkillType.Guns:
-                    factor -= character.DerivedStats.Skills[SkillType.GunsAccuracy] / 200f;
-                    factor -= character.DerivedStats.Stats[StatType.Dexterity] / 100f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.GunsAccuracy] / 200f;
+                    factor -= character.DerivedStats.Stats[(int)StatType.Dexterity] / 100f;
                     break;
             }
 
@@ -430,12 +430,12 @@ namespace CommonCore.RpgGame.Rpg
             switch ((SkillType)itemModel.SkillType)
             {
                 case SkillType.Archery:
-                    factor += character.DerivedStats.Skills[SkillType.ArcherySteady] / 100f;
-                    factor += character.DerivedStats.Stats[StatType.Dexterity] / 50f;
+                    factor += character.DerivedStats.Skills[(int)SkillType.ArcherySteady] / 100f;
+                    factor += character.DerivedStats.Stats[(int)StatType.Dexterity] / 50f;
                     break;
                 case SkillType.Guns:
-                    factor += character.DerivedStats.Skills[SkillType.GunsAccuracy] / 200f;
-                    factor += character.DerivedStats.Stats[StatType.Dexterity] / 50f;
+                    factor += character.DerivedStats.Skills[(int)SkillType.GunsAccuracy] / 200f;
+                    factor += character.DerivedStats.Stats[(int)StatType.Dexterity] / 50f;
                     break;
             }
 
@@ -450,22 +450,22 @@ namespace CommonCore.RpgGame.Rpg
             switch ((SkillType)itemModel.SkillType)
             {
                 case SkillType.Melee:
-                    factor += character.DerivedStats.Stats[StatType.Resilience] / 20f;
-                    factor += character.DerivedStats.Stats[StatType.Dexterity] / 30f;
-                    factor += character.DerivedStats.Skills[SkillType.MeleeBrawn] / 100f;
-                    factor += character.DerivedStats.Skills[SkillType.Melee] / 200f;
+                    factor += character.DerivedStats.Stats[(int)StatType.Resilience] / 20f;
+                    factor += character.DerivedStats.Stats[(int)StatType.Dexterity] / 30f;
+                    factor += character.DerivedStats.Skills[(int)SkillType.MeleeBrawn] / 100f;
+                    factor += character.DerivedStats.Skills[(int)SkillType.Melee] / 200f;
                     if (!itemModel.CheckFlag(ItemFlag.WeaponNeverRandomize))
                     {
-                        float invPrecision = Mathf.Clamp(100f - character.DerivedStats.Skills[SkillType.MeleePrecision], 0, 100f);
+                        float invPrecision = Mathf.Clamp(100f - character.DerivedStats.Skills[(int)SkillType.MeleePrecision], 0, 100f);
                         float precisionMax = invPrecision / 200f;
                         float randomFactor = UnityEngine.Random.Range(0, precisionMax);
                         factor -= randomFactor;
                     }
                     break;
                 case SkillType.Archery:
-                    factor += character.DerivedStats.Skills[SkillType.ArcheryDraw] / 100f;
-                    factor += character.DerivedStats.Stats[StatType.Resilience] / 50f;
-                    factor += character.DerivedStats.Stats[StatType.Dexterity] / 50f;
+                    factor += character.DerivedStats.Skills[(int)SkillType.ArcheryDraw] / 100f;
+                    factor += character.DerivedStats.Stats[(int)StatType.Resilience] / 50f;
+                    factor += character.DerivedStats.Stats[(int)StatType.Dexterity] / 50f;
                     break;
             }
 
@@ -480,14 +480,14 @@ namespace CommonCore.RpgGame.Rpg
             switch ((SkillType)itemModel.SkillType)
             {
                 case SkillType.Melee:
-                    factor -= character.DerivedStats.Skills[SkillType.MeleePrecision] / 100f;
-                    factor -= character.DerivedStats.Stats[StatType.Dexterity] / 100f;
-                    factor -= character.DerivedStats.Stats[StatType.Resilience] / 100f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.MeleePrecision] / 100f;
+                    factor -= character.DerivedStats.Stats[(int)StatType.Dexterity] / 100f;
+                    factor -= character.DerivedStats.Stats[(int)StatType.Resilience] / 100f;
                     break;
                 case SkillType.Archery:
-                    factor -= character.DerivedStats.Skills[SkillType.Archery] / 200f;
-                    factor -= character.DerivedStats.Skills[SkillType.ArcheryDraw] / 100f;
-                    factor -= character.DerivedStats.Stats[StatType.Dexterity] / 100f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.Archery] / 200f;
+                    factor -= character.DerivedStats.Skills[(int)SkillType.ArcheryDraw] / 100f;
+                    factor -= character.DerivedStats.Stats[(int)StatType.Dexterity] / 100f;
                     break;
             }
